Move utility scripts out of src
They aren't under test (they kind of are test code) and may be removed
eventually anyway.
